backend/python-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
librosa==0.10.1
numpy==1.24.3
scipy==1.11.1
sounddevice==0.4.6
pydub==0.25.1
redis==4.6.0
pydantic==2.1.1
python-multipart==0.0.6
aiofiles==23.1.0
httpx==0.24.1
python-dotenv==1.0.0
loguru==0.7.0
asyncio-mqtt==0.13.0
websockets==11.0.3
